==> 01_credit risk modeling preparation.ipynb <==
# Data Preparation

import numpy as np
import pandas as pd

## Load data

load_backup_data = pd.read_csv("../loan_data_2007_2014.csv")

loan_data = load_backup_data.copy()

# Delete a single column by name
loan_data = loan_data.drop('Unnamed: 0', axis=1)
loan_data

## Explore

loan_data 

loan_data.info()

# General Pre processsing

## 1. Changing continous variables from object/string to int

### Variable 1. emp_length

loan_data["emp_length"].unique()

!Note: Always performing string manipulation in order. 

For example for the column 'emp_lenght' performing regex'\D' will make both the values '< 1 year' and '1 year' same as 1, Which we are trying to avoid here.

loan_data['emp_length_int'] = loan_data['emp_length'].replace('< 1 year', '0')

loan_data['emp_length_int'].unique()

loan_data['emp_length_int'] = loan_data['emp_length_int'].fillna('0')

loan_data['emp_length_int'].unique()

# The following regex r'\D' replaces any non digit characters with an empty string
loan_data['emp_length_int'] = loan_data['emp_length_int'].replace(r'\D', '', regex = True)

loan_data['emp_length_int'].unique()

loan_data['emp_length_int'] = pd.to_numeric(loan_data['emp_length_int'])

type(loan_data['emp_length_int'][0])

### Variable 2. term

loan_data['term'].unique()

loan_data['term_int'] = loan_data['term'].replace(r'\D', '', regex = True)

loan_data['term_int'].unique()

loan_data['term_int'] = pd.to_numeric(loan_data['term_int'])

type(loan_data['term_int'][0])

## 2. Changing string variables to date variables

loan_data['earliest_cr_line'].head()

# Creating function for handling date conversion from string to datetime 
# (also handling dates before 2000 i.e 1960 for example, to avoid being labbeled as 2060)

from datetime import datetime


def convert_to_datetime(date_str):
    if pd.isnull(date_str):  # Handle missing values
        return None

    date_parts = str(date_str).split('-')
    if len(date_parts) != 2:
        return None

    month_abbr, year = date_parts
    month = datetime.strptime(month_abbr, '%b').month

    if int(year) >= 0 and int(year) <= 21:
        year = str(int(year) + 2000)
    else:
        year = str(int(year) + 1900)

    return datetime.strptime(f'01-{month}-{year}', '%d-%m-%Y')

loan_data['earliest_cr_line'] = loan_data['earliest_cr_line'].apply(convert_to_datetime)


loan_data['earliest_cr_line'].head()

loan_data['mnths_since_earliest_cr_line'] = round(pd.to_numeric((pd.to_datetime('2017-12-01') - loan_data['earliest_cr_line']) / np.timedelta64(1,"M")))

loan_data['mnths_since_earliest_cr_line'].head()

loan_data['mnths_since_earliest_cr_line'].describe()

loan_data['issue_d'] 

loan_data['issue_date'] = loan_data['issue_d'].apply(convert_to_datetime)
loan_data['issue_date'] 

loan_data['mths_since_issue_d'] = round(pd.to_numeric((pd.to_datetime('2017-12-01') - loan_data['issue_date']) / np.timedelta64(1, 'M')))
loan_data['mths_since_issue_d'].describe()

# Preprocessing Discrete Variables

Creating dummy variables for discrete variables

loan_data['grade'].unique()

pd.get_dummies(loan_data['grade'], prefix = 'Grade', prefix_sep=': ')

loan_data_dummies = [
                        pd.get_dummies(loan_data['grade'], prefix = 'grade', prefix_sep=': '), 
                        pd.get_dummies(loan_data['sub_grade'], prefix = 'sub_grade', prefix_sep=': '), 
                        pd.get_dummies(loan_data['home_ownership'], prefix = 'home_ownership', prefix_sep=': '),
                        pd.get_dummies(loan_data['verification_status'], prefix = 'verification_status', prefix_sep=': '),
                        pd.get_dummies(loan_data['loan_status'], prefix = 'loan_status', prefix_sep=': '),
                        pd.get_dummies(loan_data['purpose'], prefix = 'purpose', prefix_sep=': '),
                        pd.get_dummies(loan_data['addr_state'], prefix = 'addr_state', prefix_sep=': '),
                        pd.get_dummies(loan_data['initial_list_status'], prefix = 'initial_list_status', prefix_sep=': ')
                    ]

converting loan_data_dummies to a dataframe

loan_data_dummies = pd.concat(loan_data_dummies, axis = 1)

type(loan_data_dummies)

loan_data_dummies

### Combining dummy variable data with original dataframe i.e loan_data

# axis = 0 , refers to rows (observations);
# axis = 1, refers to columns (variables)
## We will join the tables with columns
loan_data = pd.concat([loan_data, loan_data_dummies], axis = 1)

loan_data

# Check for missing values and clean

# display only the column names with a number of missing values greater than 0,
missing_values_sum = loan_data.isnull().sum()
missing_values_sum[missing_values_sum > 0]

### Variables needed for Analysis
1. annual_inc
2. delinq_2yrs
3. inq_last_6mnths
4. open_acc
5. pub_rec
6. total_acc
7. acc_now_delinq
8. total_rev_hi_lim
9. emp_length_int
10. mths_since_earliest_cr_line

treating missing values for needed variables

loan_data['total_rev_hi_lim'].isnull().sum()

## total_rev_hi_lim

Here we assume that if we don't have revolving limit then it is equal to funded amount of the loan

loan_data['total_rev_hi_lim'].fillna(loan_data['funded_amnt'], inplace = True)

loan_data['total_rev_hi_lim'].isnull().sum()

## annual_inc

For annual income (‘annual_inc’), replace the missing values with the mean of the non-missing values of the variable.

loan_data['annual_inc'].fillna(loan_data['annual_inc'].mean(), inplace=True)

loan_data['annual_inc'].isnull().sum()

## For all of the following variables, replace the missing values with zeroes
‘mths_since_earliest_cr_line’

‘acc_now_delinq’

‘total_acc’

‘pub_rec’

‘open_acc’

‘inq_last_6mths’

‘delinq_2yrs’

‘emp_length_int’


variables = ['mnths_since_earliest_cr_line', 'acc_now_delinq', 'total_acc', 'pub_rec', 'open_acc', 'inq_last_6mths', 'delinq_2yrs', 'emp_length_int']
loan_data[variables] = loan_data[variables].fillna(0)

loan_data[variables].isnull().sum()

loan_data.to_pickle('../loan_data.pkl')
