==> 02_default_definition.ipynb <==
import numpy as np
import pandas as pd

loan_data = pd.read_pickle("../loan_data.pkl")
loan_data

## Dependent Variable (Good/Bad Defintiion). Defaulted and Non-Defaulted Accounts

loan_data['loan_status'].unique()

loan_data['loan_status'].value_counts()

seeing the numbers in proportions

loan_data['loan_status'].value_counts() / loan_data['loan_status'].count()

## Defining the Default Definition

default_list = ['Charged Off', 'Default' ,'Does not meet the credit policy. Status:Charged Off', 'Late (31-120 days)' ]

# We create a new variable that has the value of '0' if a condition is met, and the value of '1' if it is not met.
# So here a value of '0' states bad loan i.e 'Default' and 1 is good loan.
loan_data['good_bad'] = np.where(loan_data['loan_status'].isin(default_list), 0, 1)

loan_data['good_bad']

loan_data.head()

loan_data.to_pickle("../loan_data_defaultDefinition.pkl")
